cast rates to be muxed to uint32_t instead of uint16_t

+++ libraries/AC_Planck/AC_Planck.h
@@ -66,7 +66,7 @@ public:
   //Get a position, velocity, yaw command
   bool get_posvel_cmd(Location &loc, Vector3f &vel_cms, float &yaw_cd, bool &is_yaw_rate);
 
-  uint32_t mux_rates(const float rate_up, const float rate_down){return (uint16_t(rate_up) << 16) | uint16_t(rate_down);};
+  uint32_t mux_rates(const float rate_up, const float rate_down){return (uint32_t(rate_up) << 16) | uint32_t(rate_down);};
 
 private:
